reformat, add warrior, magician handler

+++ commons/src/main/java/com/msemu/commons/rmi/model/ChannelInfo.java
@@ -36,7 +36,7 @@ public class ChannelInfo implements Serializable {
 
     @Getter
     @Setter
-    private short port;
+    private int port;
 
     public int getChannelLoading() {
         if (getMaxConnectedClients() > 0)
